+++ CHANGELOG.md
@@ -16,8 +16,8 @@
 * 0.61.0  Implementation of bug fixes across bash qx_utilities and pipeline restructure back-compatibility
 * 0.60.1  Full support for hcp_suffix in run_qc.
 * 0.60.0  Renamed all subject related parameters to session. Pipeline architecture restructure.
-* 0.50.11 Updated ColeAnticevicNetPartition with files from most recent release (v1.1.4) of the parcellation.
-* 0.50.10 Updated ColeAnticevicNetPartition with files from most recent release of the parcellation.
+* 0.50.11 Updated cole_anticevic_net_partition with files from most recent release (v1.1.4) of the parcellation.
+* 0.50.10 Updated cole_anticevic_net_partition with files from most recent release of the parcellation.
 * 0.50.9  Added montage templates to the repository.
 * 0.50.8  Consistent jobname in scheduling between qunex and qunex_container.
 * 0.50.7  License and README updates.